captcha_api/requirements.txt
uvicorn==0.13.4
fastapi==0.63.0
numpy==1.18.1
loguru==0.4.1
python-multipart==0.0.5
tensorflow==2.3.2
Keras==2.4.3
opencv-python==4.5.1.48
Pillow==8.1.1
motor==2.3.1
protobuf==3.20.*